apps/services/notification-service/requirements.txt
# SAHOOL Notification Service Dependencies
# متطلبات خدمة الإشعارات
# Updated: December 2025

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3

# HTTP Client
httpx==0.28.1

# Firebase Cloud Messaging (for push notifications)
firebase-admin==6.6.0

# Redis for distributed notification queue
redis==5.2.1

# APScheduler for scheduled notifications
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1

# Database (PostgreSQL with Tortoise ORM)
asyncpg==0.30.0
tortoise-orm==0.21.7
aerich==0.7.2
